MakeCert is Depreciated....please update doc ASAP
[9:56 AM] Rob Sullivan
MakeCert is deprecated. This means that this tool could be removed at any point. Any certificates that you already generated using MakeCert won't be affected when MakeCert is no longer available. MakeCert is only used to generate the certificates, not as a validating mechanism.

[9:57 AM] Rob Sullivan
MakeCert - Win32 apps | Microsoft Docs
MakeCert - Win32 apps
Creates an X.509 certificate, signed by the test root key or other specified key, that binds your name to the public part of the key pair. The certificate is saved to a file, a system certificate s...